requirements.txt
# Function dependencies, for example:
# package>=version
firebase-admin==5.1.0
python-dotenv==0.19.2